# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DARKSOL Inference API",
    "version": "0.2.1",
    "description": "Local inference server with OpenAI-compatible endpoints and static shell assets."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:11435"
    }
  ],
  "paths": {
    "/health": {
      "get": {
        "summary": "Service health check",
        "responses": {
          "200": {
            "description": "Service status payload"
          }
        }
      }
    },
    "/web/{filePath}": {
      "get": {
        "summary": "Serve local web shell files (loopback only)",
        "responses": {
          "200": {
            "description": "Static web file"
          },
          "403": {
            "description": "Remote clients are blocked"
          },
          "404": {
            "description": "File not found"
          }
        },
        "parameters": [
          {
            "name": "filePath",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/assets/{filePath}": {
      "get": {
        "summary": "Serve local asset files (loopback only)",
        "responses": {
          "200": {
            "description": "Static asset file"
          },
          "403": {
            "description": "Remote clients are blocked"
          },
          "404": {
            "description": "File not found"
          }
        },
        "parameters": [
          {
            "name": "filePath",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/v1/models": {
      "get": {
        "summary": "List installed models in OpenAI list format (DARKSOL + Ollama when enabled)",
        "responses": {
          "200": {
            "description": "Model list"
          },
          "401": {
            "description": "Missing or invalid API key"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/v1/ollama/models": {
      "get": {
        "summary": "List Ollama local models",
        "responses": {
          "200": {
            "description": "Model list"
          },
          "400": {
            "description": "Ollama provider disabled"
          },
          "401": {
            "description": "Missing or invalid API key"
          },
          "502": {
            "description": "Ollama unavailable"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/v1/models/pull": {
      "post": {
        "summary": "Download and register a model from HuggingFace",
        "responses": {
          "200": {
            "description": "Model pull result"
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Missing or invalid API key"
          },
          "502": {
            "description": "Upstream pull failure"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/v1/models/import-ollama": {
      "post": {
        "summary": "Import an Ollama-local model into Darksol model registry",
        "responses": {
          "200": {
            "description": "Model import result"
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Missing or invalid API key"
          },
          "404": {
            "description": "Model not found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/v1/directory/models": {
      "get": {
        "summary": "Search HuggingFace model directory",
        "responses": {
          "200": {
            "description": "Directory search result"
          },
          "400": {
            "description": "Invalid query parameters"
          },
          "401": {
            "description": "Missing or invalid API key"
          },
          "502": {
            "description": "Upstream directory unavailable"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100
            }
          },
          {
            "name": "task",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "trending",
                "popular",
                "downloads",
                "recent",
                "likes"
              ]
            }
          },
          {
            "name": "hardware_aware",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "fit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "recommended",
                "will_fit",
                "might_fit",
                "any"
              ]
            }
          }
        ]
      }
    },
    "/v1/bankr/health": {
      "get": {
        "summary": "Bankr gateway config status",
        "responses": {
          "200": {
            "description": "Bankr gateway health payload"
          },
          "401": {
            "description": "Missing or invalid API key"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/v1/app/meta": {
      "get": {
        "summary": "App bootstrap metadata and route inventory",
        "responses": {
          "200": {
            "description": "Metadata payload"
          },
          "401": {
            "description": "Missing or invalid API key"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/v1/app/usage": {
      "get": {
        "summary": "Local inference usage and cost totals",
        "responses": {
          "200": {
            "description": "Usage payload"
          },
          "401": {
            "description": "Missing or invalid API key"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/v1/runtime/status": {
      "get": {
        "summary": "Darksol Engine runtime status plus keep-warm state",
        "responses": {
          "200": {
            "description": "Runtime status payload"
          },
          "401": {
            "description": "Missing or invalid API key"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/v1/runtime/start": {
      "post": {
        "summary": "Start managed Darksol Engine runtime",
        "responses": {
          "200": {
            "description": "Runtime start result"
          },
          "401": {
            "description": "Missing or invalid API key"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/v1/runtime/stop": {
      "post": {
        "summary": "Stop managed Darksol Engine runtime",
        "responses": {
          "200": {
            "description": "Runtime stop result"
          },
          "401": {
            "description": "Missing or invalid API key"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/v1/runtime/restart": {
      "post": {
        "summary": "Restart managed Darksol Engine runtime",
        "responses": {
          "200": {
            "description": "Runtime restart result"
          },
          "401": {
            "description": "Missing or invalid API key"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/v1/runtime/ports": {
      "get": {
        "summary": "Check if a host/port is available for runtime binding",
        "responses": {
          "200": {
            "description": "Port availability payload"
          },
          "401": {
            "description": "Missing or invalid API key"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/v1/runtime/ports/find": {
      "post": {
        "summary": "Find a free port for runtime binding",
        "responses": {
          "200": {
            "description": "Found free port"
          },
          "400": {
            "description": "Invalid start port"
          },
          "401": {
            "description": "Missing or invalid API key"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/v1/runtime/config": {
      "post": {
        "summary": "Update runtime network config (host/port)",
        "responses": {
          "200": {
            "description": "Runtime config update result"
          },
          "400": {
            "description": "Invalid host/port"
          },
          "401": {
            "description": "Missing or invalid API key"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/v1/runtime/keepwarm": {
      "get": {
        "summary": "Get Darksol Engine keep-warm configuration",
        "responses": {
          "200": {
            "description": "Keep-warm configuration payload"
          },
          "401": {
            "description": "Missing or invalid API key"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "summary": "Update Darksol Engine keep-warm configuration",
        "responses": {
          "200": {
            "description": "Keep-warm update result"
          },
          "401": {
            "description": "Missing or invalid API key"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/v1/mcp/servers": {
      "get": {
        "summary": "List configured MCP servers",
        "responses": {
          "200": {
            "description": "MCP server list"
          },
          "401": {
            "description": "Missing or invalid API key"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/v1/mcp/servers/{name}/enable": {
      "post": {
        "summary": "Enable a configured MCP server",
        "responses": {
          "200": {
            "description": "MCP server enabled"
          },
          "401": {
            "description": "Missing or invalid API key"
          },
          "404": {
            "description": "MCP server not found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/v1/mcp/servers/{name}/disable": {
      "post": {
        "summary": "Disable a configured MCP server",
        "responses": {
          "200": {
            "description": "MCP server disabled"
          },
          "401": {
            "description": "Missing or invalid API key"
          },
          "404": {
            "description": "MCP server not found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/v1/chat/completions": {
      "post": {
        "summary": "OpenAI-compatible chat completions",
        "responses": {
          "200": {
            "description": "Chat completion result"
          },
          "400": {
            "description": "Request validation error"
          },
          "401": {
            "description": "Missing or invalid API key"
          },
          "404": {
            "description": "Model not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/v1/completions": {
      "post": {
        "summary": "OpenAI-compatible text completions",
        "responses": {
          "200": {
            "description": "Text completion result"
          },
          "400": {
            "description": "Request validation error"
          },
          "401": {
            "description": "Missing or invalid API key"
          },
          "404": {
            "description": "Model not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/v1/embeddings": {
      "post": {
        "summary": "OpenAI-compatible embeddings",
        "responses": {
          "200": {
            "description": "Embedding result"
          },
          "400": {
            "description": "Request validation error"
          },
          "401": {
            "description": "Missing or invalid API key"
          },
          "404": {
            "description": "Model not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API Key"
      }
    }
  }
}